Chessito - application restructuring
