--- src/lichess/lichess.js
@@ -1,12 +1,0 @@
-import { login } from "./lichess-api.js";
-
-export default class Lichess {
-
-    constructor(credentials) {
-        this.credentials = credentials;
-    }
-
-    login() {
-        return login(this.credentials);
-    }
-}
