+++ src/models/lichess.js
@@ -1,0 +1,8 @@
+export function login(credentials) {
+    console.log('login');
+    console.log(credentials);
+}
+
+export function getLiveGames() {
+    console.log('getLiveGames');
+}
