+++ src/services/lichess.js
@@ -1,0 +1,16 @@
+import { login, getLiveGames } from "../models/lichess.js";
+
+export default class LichessService {
+
+    constructor(credentials) {
+        this.credentials = credentials;
+    }
+
+    login() {
+        return login(this.credentials);
+    }
+
+    getLiveGames() {
+        return getLiveGames();
+    }
+}
